Removed Neural.NET API and reenabled local network
