--- MNISTApp/MNISTApp.Windows/MainPage.xaml
@@ -1,26 +1,0 @@
-﻿<Page
-    x:Class="MNISTApp.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:MNISTApp"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <Grid Background="Black">
-        <Border BorderThickness="2" BorderBrush="White" Background="Black" HorizontalAlignment="Center"  VerticalAlignment="Center" Width="380" Height="380">
-            <Canvas x:Name="DrawingCanvas" PointerPressed="DrawingCanvas_PointerPressed" PointerMoved="DrawingCanvas_PointerMoved" PointerReleased="DrawingCanvas_PointerReleased">
-                <Canvas.Background>
-                    <SolidColorBrush Color="Black" />
-                </Canvas.Background>
-            </Canvas>
-        </Border>
-    </Grid>
-
-    <Page.BottomAppBar>
-        <CommandBar>
-            <AppBarButton Icon="ContactInfo"  Label="About" Click="AppBarButton_Click" />
-        </CommandBar>
-    </Page.BottomAppBar>
-</Page>
