--- MNISTApp/MNISTApp.Windows/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FSharp.Core" version="3.1.2.1" targetFramework="win81" />
-  <package id="MathNet.Numerics" version="3.5.0" targetFramework="win81" />
-</packages>
